pyproject.toml
[tool.poetry]
name = "stac-dataloader"
version = "0.3.0"
description = "Dataset interfaces for loading STAC references for Machine Learning pipelines."
authors = ["Francis Charette-Migneault <francis.charette.migneault@crim.ca>"]
license = "MIT"
readme = "README.md"
packages = [{include = "stac_dataloader"}]

[tool.poetry.dependencies]
python = "^3.8"
rasterio = "^1.3.6"
torchgeo = "^0.4.1"
typing-extensions = "^4.5.0"
pystac = "^1.7.3"
pystac-client = "^0.6.1"
torchdata = "^0.6.0"
zen3geo = "^0.6.0"
planetary-computer = "^0.5.1"
xbatcher = "^0.3.0"
stackstac = "^0.4.3"
graphviz = "^0.20.1"
leafmap = "^0.20.3"
jupyterlab = "<4"
jupyter = "^1.0.0"
nbdime = "^3.2.1"
tornado = "6.1"
shapely = ">=2"
fastapi = ">=0.80"
websockets = "^11.0.3"
dateutils = "^0.6.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
bump2version = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
